// src/fsm/types.ts
export type LockState =
  | 'LOCKED'
  | 'UNLOCKED'
  | 'ATTEMPT'
  | 'TIMEOUT'
  | 'SET_PASSWORD'
  | 'CONFIRM_PASSWORD';

export type LockEvent =
  | 'START_ATTEMPT'
  | 'CORRECT_PASSWORD'
  | 'WRONG_PASSWORD'
  | 'MAX_ATTEMPTS'
  | 'TIMEOUT_COMPLETE'
  | 'START_SET_PASSWORD'
  | 'PASSWORD_SUBMITTED'
  | 'CONFIRM_PASSWORD_MATCH'
  | 'CONFIRM_PASSWORD_MISMATCH'
  | 'MANUAL_LOCK';